+++ internal/clients/clientimpl/localmatcher/zip.go
@@ -1,4 +1,4 @@
-package clientimpl
+package localmatcher
 
 import (
 	"archive/zip"
