+++ internal/clients/clientimpl/localmatcher/zip_test.go
@@ -1,4 +1,4 @@
-package clientimpl_test
+package localmatcher_test
 
 import (
 	"archive/zip"
@@ -16,7 +16,7 @@ import (
 	"sort"
 	"testing"
 
-	"github.com/google/osv-scanner/internal/clients/clientimpl"
+	"github.com/google/osv-scanner/internal/clients/clientimpl/localmatcher"
 	"github.com/google/osv-scanner/internal/testutility"
 	"github.com/google/osv-scanner/pkg/models"
 )
@@ -145,10 +145,10 @@ func TestNewZippedDB_Offline_WithoutCache(t *testing.T) {
 		t.Errorf("a server request was made when running offline")
 	})
 
-	_, err := clientimpl.NewZippedDB(testDir, "my-db", ts.URL, true)
+	_, err := localmatcher.NewZippedDB(testDir, "my-db", ts.URL, true)
 
-	if !errors.Is(err, clientimpl.ErrOfflineDatabaseNotFound) {
-		t.Errorf("expected \"%v\" error but got \"%v\"", clientimpl.ErrOfflineDatabaseNotFound, err)
+	if !errors.Is(err, localmatcher.ErrOfflineDatabaseNotFound) {
+		t.Errorf("expected \"%v\" error but got \"%v\"", localmatcher.ErrOfflineDatabaseNotFound, err)
 	}
 }
 
@@ -177,7 +177,7 @@ func TestNewZippedDB_Offline_WithCache(t *testing.T) {
 		"GHSA-5.json": {ID: "GHSA-5"},
 	}))
 
-	db, err := clientimpl.NewZippedDB(testDir, "my-db", ts.URL, true)
+	db, err := localmatcher.NewZippedDB(testDir, "my-db", ts.URL, true)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
@@ -195,7 +195,7 @@ func TestNewZippedDB_BadZip(t *testing.T) {
 		_, _ = w.Write([]byte("this is not a zip"))
 	})
 
-	_, err := clientimpl.NewZippedDB(testDir, "my-db", ts.URL, false)
+	_, err := localmatcher.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err == nil {
 		t.Errorf("expected an error but did not get one")
@@ -207,7 +207,7 @@ func TestNewZippedDB_UnsupportedProtocol(t *testing.T) {
 
 	testDir := testutility.CreateTestDir(t)
 
-	_, err := clientimpl.NewZippedDB(testDir, "my-db", "file://hello-world", false)
+	_, err := localmatcher.NewZippedDB(testDir, "my-db", "file://hello-world", false)
 
 	if err == nil {
 		t.Errorf("expected an error but did not get one")
@@ -237,7 +237,7 @@ func TestNewZippedDB_Online_WithoutCache(t *testing.T) {
 		})
 	})
 
-	db, err := clientimpl.NewZippedDB(testDir, "my-db", ts.URL, false)
+	db, err := localmatcher.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
@@ -269,7 +269,7 @@ func TestNewZippedDB_Online_WithoutCacheAndNoHashHeader(t *testing.T) {
 		}))
 	})
 
-	db, err := clientimpl.NewZippedDB(testDir, "my-db", ts.URL, false)
+	db, err := localmatcher.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
@@ -307,7 +307,7 @@ func TestNewZippedDB_Online_WithSameCache(t *testing.T) {
 
 	cacheWrite(t, determineStoredAtPath(testDir, "my-db"), cache)
 
-	db, err := clientimpl.NewZippedDB(testDir, "my-db", ts.URL, false)
+	db, err := localmatcher.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
@@ -345,7 +345,7 @@ func TestNewZippedDB_Online_WithDifferentCache(t *testing.T) {
 		"GHSA-3.json": {ID: "GHSA-3"},
 	}))
 
-	db, err := clientimpl.NewZippedDB(testDir, "my-db", ts.URL, false)
+	db, err := localmatcher.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
@@ -375,7 +375,7 @@ func TestNewZippedDB_Online_WithCacheButNoHashHeader(t *testing.T) {
 		"GHSA-3.json": {ID: "GHSA-3"},
 	}))
 
-	_, err := clientimpl.NewZippedDB(testDir, "my-db", ts.URL, false)
+	_, err := localmatcher.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err == nil {
 		t.Errorf("expected an error but did not get one")
@@ -403,7 +403,7 @@ func TestNewZippedDB_Online_WithBadCache(t *testing.T) {
 
 	cacheWriteBad(t, determineStoredAtPath(testDir, "my-db"), "this is not json!")
 
-	db, err := clientimpl.NewZippedDB(testDir, "my-db", ts.URL, false)
+	db, err := localmatcher.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
@@ -429,7 +429,7 @@ func TestNewZippedDB_FileChecks(t *testing.T) {
 		})
 	})
 
-	db, err := clientimpl.NewZippedDB(testDir, "my-db", ts.URL, false)
+	db, err := localmatcher.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
