+++ internal/clients/clientimpl/osvmatcher/osvmatcher.go
@@ -1,4 +1,4 @@
-package clientimpl
+package osvmatcher
 
 import (
 	"context"
